Refactor geolocation function to controller file
